Add validation step to test_ote_training.py for stages that were run from dependency chain earlier
It is possible that a stage is run from a dependency chain (for some
other stage) and later it is run not from dependency chain.
In this case the stage's action will be run in the first run only, but
validation will be run each time when the stage is run not from a
dependency chain.